+++ dawanow-api/src/main/java/org/dawanow/dawanowapi/services/PharmacistService.java
@@ -1,15 +1,15 @@
 package org.dawanow.dawanowapi.services;
 
 import org.dawanow.dawanowapi.dto.NearestProviderDTO;
-import org.dawanow.dawanowapi.dto.PharmacyRegisterRequestDTO;
-import org.dawanow.dawanowapi.dto.PharmacyRegisterResponseDTO;
+import org.dawanow.dawanowapi.dto.pharmacy.PharmacyRegisterRequestDTO;
+import org.dawanow.dawanowapi.dto.pharmacy.PharmacyRegisterResponseDTO;
 import org.dawanow.dawanowapi.models.User;
 
 import java.util.List;
 
 public interface PharmacistService {
 
-    List<NearestProviderDTO> getNearestProviders(int pharmacyId, float radius);
+    List<NearestProviderDTO> getNearestProviders(int pharmacyId, double radius);
     PharmacyRegisterResponseDTO registerPharmacist(User user, PharmacyRegisterRequestDTO requestDTO);
 
 }
